Move the template loading logic to its own module, and migrate to TypeScript
